refactor : 퀴즈 결과 페이지 더미 데이터 위치 변경경
